Discovery:  Do full local discovery when changing the ignore files config
Since we are now skipping the full local discovery, we also need to
call the method to force a full local discovery while we were making
full remote discovery, for the same reason.

Same applies when applying a selective sync: Normally there should not
be any file or directory with the same same of a blacklisted folder, but
if there is one, it was ignored and now need to be taken in account.

Also make the F6 shortcut do a full local discovery as well.