Add an Antora playbook file
It's come up discussion around the migration to Antora, that the client
docs should be able to be built standalone, and not just as a part of
the larger documentation set. By adding an Antora Playbook file
(site.yml), the client docs can be built both standalone and as part of
the larger ownCloud documentation site.

See https://github.com/owncloud/client/pull/6785#issuecomment-425391756