Database: future-proof to allow downgrade from future version
The 2.6 client will introduce an index that depends on a custom
function.
This causes insersion onto the metadata database to fail if the database
it opened with older version of the client. (Which will cause the client
to abort)

Delete this index in this version in order to allow downgrade.

In addition, allow to load folder with version '2', but still write
version '1' in the config file

Of course this commit need not to be in the 2.6 release