Accept all ignored deletes
If the local permissions are not correct we perform a prohibited move
The move fails with 403 and creates a blacklist entry, but we might get the correct permissions for that file.

In the next sync the we don't retry the move as we have the correct permissions now so we want to restore the original file.
The original file however is now on the blacklist we mark the restoration as ignore and end up in findAndCancelDeletedJob

Fixes: #8837