Mac: Move launch-on-login to use launchd
Previously, deprecated API was used to tell macOS to launch the GUI when
the user logs in. This has now been moved to launchd, which uses the
contents of the file
~/Library/LaunchAgents/com.owncloud.desktopclient.plist to determine
what to do. This comes with the added benefit that wel now tell launchd
to relaunch the client when it crashes.

Fixes: #9037