Fix FakeGetReply for non-existing files
The old code didn't take into account the case where a file was removed
on the server. It just Q_ASSERT-ed that it should always exist. However,
the CI does release builds, so the assert never triggered.

The problem is that in TestSyncVirtualFiles::testVirtualFileDownload the
file "A/a3" *is* removed from the server while a download is scheduled.

The fix is to handle this case separately as a valid response.