Load log file count limit properly from config file
This commit makes sure that when starting the application, the value is populated from the config file. Previously, the value was only changed when the user set a different value in the log settings dialog.

The static initialization was removed from the constructor to prevent a cyclic dependency between the config file and logger classes (the latter uses the former to populate the value mentioned above).