Don't start a timer per changed file
The client used to emit a file changed event for every single change.
In order to prevent pointless syncs it delayed the sync by 2s.
This created a QTimer for every singe change.

Since 5.0 we gather changes for 10s and only then handle the changes, so no delay should be needed anymore.