Signal AppKit that we are fine with SecureCoding
Since macOS 12 there is a feature to do object de-serialization more
secure (secure coding). Starting with with macOS 14, AppKit starts to
emit a warning (on the command-line) when an application does not
support this feature.

We, nor Qt, use object (de)serialization with `NSCoder`, so we can
safely signal back to AppKit that we are fine with secure coding.

This surpresses the AppKit warning:

Secure coding is not enabled for restorable state! Enable secure coding by implementing NSApplicationDelegate.applicationSupportsSecureRestorableState: and returning YES.

Fixes: #11191