fix crash when Url has a non-empty query on adding account (#12118)
* fix crash when adding an accountState to the AccountManager

slot to saveAccount was connected before AccountState was effectively added. saveAccount uses the AccountStatePtr so it was null in some cases and crashed.

I added checks for null on the account state in these areas and also re-ordered the behavior in AccountManager::addAccountState

also renamed AccountManager::account to accountState as that's actually what you get from it.

* Revert "fix crash when adding an accountState to the AccountManager"

This reverts commit cc6c0c522047370e359792c0de1b1ae1457dfd16.

* fixed crash related to QUrl having a non empty query when adding account

instead of waiting for the assert "down below" to crash the app when the QUrl has a query, check for it in the wizard handler and emit wizard evaluationFailed to handle it more gracefully.

partially fixes #7160

* fixed crash related to QUrl having a non empty query when adding account

instead of waiting for the assert "down below" to crash the app when the QUrl has a query, check for it in the wizard handler and emit wizard evaluationFailed to handle it more gracefully.

partially fixes #7160

* fix translation format