[DC-30] fix: memory leak and general cleanup of AccountSettings (#12146)
* we have a doozy of a memory leak in AccountSettings::checkConnectivity

see item #13 in the refactoring todo list

* added the new method to cleanup the existing validator, before changing QPointer to raw ptrs

* replace qptr with raw pointer

* the raw pointer checks work exactly as they should and we skip creating a new validator unless the member is actually null.

I will clean up a bit more now.

* got rid of the self-destructing aspect of ConnectionValidator and now we use normal RAII to manage this resource.
SO much healthier (especially in the AccountState impls)

also moved various connections to dedicated connectXXX methods. this not only makes the ctr much easier to read, but in future we may want to connect at a different point in time, OUTSIDE the ctr. now we can move them around super easily.

next step is obliterating the lambdas...after a nice long easter break I should have the energy for this again.

* moved handlers for NetworkInformation signals to REAL (private) functions
also refactored the very strangely implemented dialog to confirm updating the url after a server redirect. I still need to find a way to test that but coceptually this is now cleaned up A LOT
got rid of the supportsSpaces property as it had a weird notification for "supportsSpacesChanged" that made no sense, then I could not find anyone using the property anyway. Goodbye!

* just added a new todo

* fix squish tests
put the mysterious property that requires a "supportsSpacesChanged" signal back, as removing that broke the bottom area of the account settings page

* final work - move the ssl error lambda to a new function and simplify the handling of the updated certificate acceptance dialog

* a bit of cleanup

* fix: typos

* address Thomas's review comments

---------

Co-authored-by: Thomas Müller <1005065+DeepDiver1975@users.noreply.github.com>